test/reftest/vendor/blink/fast/dom/CSSStyleDeclaration/css-style-declaration-named-setter-expected.txt

PASS document.body.style.margin is "0px"
PASS document.body.style.margin = 1 is 1
PASS document.body.style.margin = badString threw exception Exception in toString().
PASS successfullyParsed is true

TEST COMPLETE
